refactoring. extract the parsing part to a new file cli-parser.ts, move the command-runner from the legacy to Harmony
